YelpDb parses json files

+++ src/main/java/ca/ece/ubc/cpen221/mp5/YelpUser.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public class YelpUser extends GenericUser{
 
 	public YelpUser(String id, String url, HashMap<String, Integer> votes, int reviewCount, String name,
-			int averageStars) {
+			double averageStars) {
 		super(id, url, votes, reviewCount, name, averageStars);
 	}
